[CLN] website_sale: move cart logic to dedicated file
Move cart logic to cart.js and modify selector to only instantiate that widget
on the cart page.

Simplify/reduce js code running on all ecommerce pages
Avoid potential side effects or conflicts with other logic (website_sale_collect/js_delete_product)
Simplify `website_sale.j` which should only hold global ecommerce logic.
