+++ addons/website_sale/static/src/js/website_sale.js
@@ -8,10 +8,10 @@ import "@website/libs/zoomodoo/zoomodoo";
 import {extraMenuUpdateCallbacks} from "@website/js/content/menu";
 import { ProductImageViewer } from "@website_sale/js/components/website_sale_image_viewer";
 import { rpc } from "@web/core/network/rpc";
-import { debounce, throttleForAnimation } from "@web/core/utils/timing";
+import { throttleForAnimation } from "@web/core/utils/timing";
 import { listenSizeChange, SIZES, utils as uiUtils } from "@web/core/ui/ui_service";
 import { isBrowserFirefox, hasTouch } from "@web/core/browser/feature_detection";
-import { Component } from "@odoo/owl";
+
 
 export const WebsiteSale = publicWidget.Widget.extend(VariantMixin, cartHandlerMixin, {
     selector: '.oe_website_sale',
@@ -19,8 +19,6 @@ export const WebsiteSale = publicWidget.Widget.extend(VariantMixin, cartHandlerM
         'change form .js_product:first input[name="add_qty"]': '_onChangeAddQuantity',
         'mouseup .js_publish': '_onMouseupPublish',
         'touchend .js_publish': '_onMouseupPublish',
-        'change .oe_cart input.js_quantity[data-product-id]': '_onChangeCartQuantity',
-        'click .oe_cart a.js_add_suggested_products': '_onClickSuggestedProduct',
         'click a.js_add_cart_json': '_onClickAddCartJSON',
         'click .a-submit': '_onClickSubmit',
         'change form.js_attributes input, form.js_attributes select': '_onChangeAttribute',
@@ -45,8 +43,6 @@ export const WebsiteSale = publicWidget.Widget.extend(VariantMixin, cartHandlerM
      */
     init: function () {
         this._super.apply(this, arguments);
-
-        this._changeCartQuantity = debounce(this._changeCartQuantity.bind(this), 500);
 
         this.isWebsite = true;
         this.filmStripStartX = 0;
@@ -198,43 +194,6 @@ export const WebsiteSale = publicWidget.Widget.extend(VariantMixin, cartHandlerM
     _changeAttribute: function (valueSelectors) {
         valueSelectors.forEach((selector) => {
             $(selector).removeClass("active").filter(":has(input:checked)").addClass("active");
-        });
-    },
-    /**
-     * @private
-     */
-    _changeCartQuantity: function ($input, value, $dom_optional, line_id, productIDs) {
-        $($dom_optional).toArray().forEach((elem) => {
-            $(elem).find('.js_quantity').text(value);
-            productIDs.push($(elem).find('span[data-product-id]').data('product-id'));
-        });
-        $input.data('update_change', true);
-
-        rpc("/shop/cart/update_json", {
-            line_id: line_id,
-            product_id: parseInt($input.data('product-id'), 10),
-            set_qty: value,
-            display: true,
-        }).then((data) => {
-            $input.data('update_change', false);
-            var check_value = parseInt($input.val() || 0, 10);
-            if (isNaN(check_value)) {
-                check_value = 1;
-            }
-            if (value !== check_value) {
-                $input.trigger('change');
-                return;
-            }
-            if (!data.cart_quantity) {
-                return window.location = '/shop/cart';
-            }
-            $input.val(data.quantity);
-            $('.js_quantity[data-line-id='+line_id+']').val(data.quantity).text(data.quantity);
-
-            wSaleUtils.updateCartNavBar(data);
-            wSaleUtils.showWarning(data.notification_info.warning);
-            // Propagating the change to the express checkout forms
-            Component.env.bus.trigger('cart_amount_changed', [data.amount, data.minor_amount]);
         });
     },
     /**
@@ -471,33 +430,6 @@ export const WebsiteSale = publicWidget.Widget.extend(VariantMixin, cartHandlerM
      */
     _onMouseupPublish: function (ev) {
         $(ev.currentTarget).parents('.thumbnail').toggleClass('disabled');
-    },
-    /**
-     * @private
-     * @param {Event} ev
-     */
-    _onChangeCartQuantity: function (ev) {
-        var $input = $(ev.currentTarget);
-        if ($input.data('update_change')) {
-            return;
-        }
-        var value = parseInt($input.val() || 0, 10);
-        if (isNaN(value)) {
-            value = 1;
-        }
-        var $dom = $input.closest('tr');
-        // var default_price = parseFloat($dom.find('.text-danger > span.oe_currency_value').text());
-        var $dom_optional = $dom.nextUntil(':not(.optional_product.info)');
-        var line_id = parseInt($input.data('line-id'), 10);
-        var productIDs = [parseInt($input.data('product-id'), 10)];
-        this._changeCartQuantity($input, value, $dom_optional, line_id, productIDs);
-    },
-    /**
-     * @private
-     * @param {Event} ev
-     */
-    _onClickSuggestedProduct: function (ev) {
-        $(ev.currentTarget).prev('input').val(1).trigger('change');
     },
     /**
      * @private
