+++ addons/website_sale/views/templates.xml
@@ -2056,7 +2056,7 @@
                 <div class="oe_structure" id="oe_structure_website_sale_cart_2"/>
             </t>
 
-            <div class="col">
+            <div class="col" id="shop_cart">
                 <h3 class="mb-4">Order overview</h3>
                 <div t-if="abandoned_proceed or access_token" class="alert alert-info mt8 mb8" role="alert"> <!-- abandoned cart choices -->
                     <t t-if="abandoned_proceed">
